+++ Solidity/27_Payable.sol
@@ -4,15 +4,13 @@ pragma solidity >=0.5.0;
 contract Payable {
     address payable public owner; // owner will be able to send eth
 
-    constructor ( ) {
+    constructor() {
         owner = payable(msg.sender);
     }
 
-    function deposit() external payable {
-        
-    }
+    function deposit() external payable {}
 
-    function getBalance() external view returns (uint) {
+    function getBalance() external view returns (uint256) {
         return address(this).balance;
     }
 }
