+++ Solidity/30_EtherWallet.sol
@@ -8,20 +8,20 @@ pragma solidity >=0.5.0;
 contract EtherWallet {
     address payable public owner;
 
-    constructor ( ) {
+    constructor() {
         owner = payable(msg.sender);
     }
 
     receive() external payable {}
 
-    function withDraw(uint _amount) external {
-        require (msg.sender == owner, "caller is not owner");
+    function withDraw(uint256 _amount) external {
+        require(msg.sender == owner, "caller is not owner");
 
         // owner.transfer(_amount); // to save gas, do this
         payable(msg.sender).transfer(_amount);
     }
 
-    function getBalance() external view returns (uint) {
+    function getBalance() external view returns (uint256) {
         return address(this).balance;
     }
 }
