+++ Solidity/31_CallOtherContracts.sol
@@ -2,43 +2,49 @@
 pragma solidity >=0.5.0;
 
 contract CallTestContracts {
-    function setX(TestContract _test, uint _x) external {
+    // function setX(address _test, uint _x) external {
+    //     TestContract(_test).setX(_x);
+    function setX(TestContract _test, uint256 _x) external {
         _test.setX(_x);
     }
 
-    function getX(TestContract _test) external view returns (uint) {
+    function getX(TestContract _test) external view returns (uint256) {
         return _test.getX();
     }
 
-    function setXandSendEther(TestContract _test, uint _x) external payable {
+    function setXandSendEther(TestContract _test, uint256 _x) external payable {
         _test.setXandReceiveEther{value: msg.value}(_x);
     }
 
-    function getXandValue(TestContract _test) external view returns (uint, uint) {
-        (uint x, uint value) =  _test.getXandValue();
+    function getXandValue(TestContract _test)
+        external
+        view
+        returns (uint256, uint256)
+    {
+        (uint256 x, uint256 value) = _test.getXandValue();
 
         return (x, value);
     }
 }
 
 contract TestContract {
-    uint public x;
-    uint public value = 123;
+    uint256 public x;
+    uint256 public value = 123;
 
-    function setX(uint _x) external {
+    function setX(uint256 _x) external {
         x = _x;
     }
 
-    function getX() external view returns (uint) {
+    function getX() external view returns (uint256) {
         return x;
     }
 
-    function setXandReceiveEther(uint _x) external payable {
+    function setXandReceiveEther(uint256 _x) external payable {
         x = _x;
         value = msg.value;
     }
 
-    function getXandValue() external view returns (uint, uint) {
+    function getXandValue() external view returns (uint256, uint256) {
         return (x, value);
     }
 }
